refactor(football): apply DRY/SRP/KISS to team and player commands
- Extract _assign_slots() module-level fn shared by both lineup paths,
  eliminating duplicated slot-fill loop and scarcity_order dict
- Move name-matching logic (find_best_match, _name_tokens, _token_substring_hit)
  from FootballTeamCommand to TheSportsDBService where SportsDBTeam lives
- Move market-value parsing (parse_market_value_millions, sum_market_values)
  from football_team command to TransfermarktService
- Extract CLUB_PREFIXES to bot/data/football_team_prefixes.py data module
- Move _FOOT_KEYS/_HEIGHT_KEYS/_OTHER_POS_KEYS/_BORN_*_KEYS from module scope
  to FootballPlayerCommand class attributes
