+++ bot/data/football_team_prefixes.py
@@ -1,0 +1,5 @@
+CLUB_PREFIXES: frozenset[str] = frozenset({
+    'fc', 'ac', 'ca', 'cd', 'cs', 'sc', 'se', 'sg', 'ec', 'cr', 'rb', 'vfb',
+    'afc', 'rcd', 'aj', 'as', 'ogc', 'ssc', 'us', 'club', 'clube', 'cf', 'sd',
+    'sv', 'sk', 'stade', 'hellas', 'sport',
+})  # fmt: skip
